Move ssl to new file
