+++ src/components/mod.rs
@@ -1,2 +1,3 @@
 pub mod banner;
+pub mod ssl_certs;
 pub mod uptime;
